+++ lib/pure/optionals.nim
@@ -16,6 +16,8 @@
 ## absence of a value is often represented by `nil`, but it is not always
 ## available, nor is it always a good solution.
 ##
+## For more advanced funcionality see `optutils module <optutils.html>`_.
+##
 ##
 ## Basic usage
 ## ===========
@@ -211,123 +213,6 @@ proc get*[T](self: var Option[T]): var T =
     raise newException(UnpackError, "Can't obtain a value from a `none`")
   return self.val
 
-proc map*[T](self: Option[T], callback: proc (input: T)) =
-  ## Applies a `callback` function to the value of the `Option`, if it has one.
-  ##
-  ## See also:
-  ## * `map proc <#map,Option[T],proc(T)_2>`_ for a version with a callback
-  ##   which returns a value
-  ## * `filter proc <#filter,Option[T],proc(T)>`_
-  runnableExamples:
-    var d = 0
-    proc saveDouble(x: int) =
-      d = 2*x
-
-    let
-      a = some(42)
-      b = none(int)
-
-    b.map(saveDouble)
-    assert d == 0
-    a.map(saveDouble)
-    assert d == 84
-
-  if self.isSome:
-    callback(self.val)
-
-proc map*[T, R](self: Option[T], callback: proc (input: T): R): Option[R] =
-  ## Applies a `callback` function to the value of the `Option` and returns an
-  ## `Option` containing the new value.
-  ##
-  ## If the `Option` is `None`, `None` of the return type of the `callback`
-  ## will be returned.
-  ##
-  ## See also:
-  ## * `flatMap proc <#flatMap,Option[A],proc(A)>`_ for a version with a
-  ##   callback which returns an `Option`
-  ## * `filter proc <#filter,Option[T],proc(T)>`_
-  runnableExamples:
-    var
-      a = some(42)
-      b = none(int)
-
-    proc isEven(x: int): bool =
-      x mod 2 == 0
-
-    assert $(a.map(isEven)) == "Some(true)"
-    assert $(b.map(isEven)) == "None[bool]"
-
-  if self.isSome:
-    some[R]( callback(self.val) )
-  else:
-    none(R)
-
-proc flatten*[A](self: Option[Option[A]]): Option[A] =
-  ## Remove one level of structure in a nested `Option`.
-  runnableExamples:
-    let a = some(some(42))
-    assert $flatten(a) == "Some(42)"
-
-  if self.isSome:
-    self.val
-  else:
-    none(A)
-
-proc flatMap*[A, B](self: Option[A], callback: proc (input: A): Option[B]): Option[B] =
-  ## Applies a `callback` function to the value of the `Option` and returns an
-  ## `Option` containing the new value.
-  ##
-  ## If the `Option` is `None`, `None` of the return type of the `callback`
-  ## will be returned.
-  ##
-  ## Similar to `map`, with the difference that the `callback` returns an
-  ## `Option`, not a raw value. This allows multiple procs with a
-  ## signature of `A -> Option[B]` to be chained together.
-  ##
-  ## See also:
-  ## * `flatten proc <#flatten,Option[Option[A]]>`_
-  ## * `filter proc <#filter,Option[T],proc(T)>`_
-  runnableExamples:
-    proc doublePositives(x: int): Option[int] =
-      if x > 0:
-        return some(2*x)
-      else:
-        return none(int)
-    let
-      a = some(42)
-      b = none(int)
-      c = some(-11)
-    assert a.flatMap(doublePositives) == some(84)
-    assert b.flatMap(doublePositives) == none(int)
-    assert c.flatMap(doublePositives) == none(int)
-
-  map(self, callback).flatten()
-
-proc filter*[T](self: Option[T], callback: proc (input: T): bool): Option[T] =
-  ## Applies a `callback` to the value of the `Option`.
-  ##
-  ## If the `callback` returns `true`, the option is returned as `Some`.
-  ## If it returns `false`, it is returned as `None`.
-  ##
-  ## See also:
-  ## * `map proc <#map,Option[T],proc(T)_2>`_
-  ## * `flatMap proc <#flatMap,Option[A],proc(A)>`_
-  runnableExamples:
-    proc isEven(x: int): bool =
-      x mod 2 == 0
-    let
-      a = some(42)
-      b = none(int)
-      c = some(-11)
-    assert a.filter(isEven) == some(42)
-    assert b.filter(isEven) == none(int)
-    assert c.filter(isEven) == none(int)
-
-  if self.isSome and not callback(self.val):
-    none(T)
-  else:
-    self
-
 proc `==`*(a, b: Option): bool =
   ## Returns `true` if both `Option`s are `None`,
   ## or if they are both `Some` and have equal values.
@@ -430,49 +315,6 @@ when isMainModule:
       check($(some("Correct")) == "Some(\"Correct\")")
       check($(stringNone) == "None[string]")
 
-    test "map with a void result":
-      var procRan = 0
-      some(123).map(proc (v: int) = procRan = v)
-      check procRan == 123
-      intNone.map(proc (v: int) = check false)
-
-    test "map":
-      check(some(123).map(proc (v: int): int = v * 2) == some(246))
-      check(intNone.map(proc (v: int): int = v * 2).isNone)
-
-    test "filter":
-      check(some(123).filter(proc (v: int): bool = v == 123) == some(123))
-      check(some(456).filter(proc (v: int): bool = v == 123).isNone)
-      check(intNone.filter(proc (v: int): bool = check false).isNone)
-
-    test "flatMap":
-      proc addOneIfNotZero(v: int): Option[int] =
-        if v != 0:
-          result = some(v + 1)
-        else:
-          result = none(int)
-
-      check(some(1).flatMap(addOneIfNotZero) == some(2))
-      check(some(0).flatMap(addOneIfNotZero) == none(int))
-      check(some(1).flatMap(addOneIfNotZero).flatMap(addOneIfNotZero) == some(3))
-
-      proc maybeToString(v: int): Option[string] =
-        if v != 0:
-          result = some($v)
-        else:
-          result = none(string)
-
-      check(some(1).flatMap(maybeToString) == some("1"))
-
-      proc maybeExclaim(v: string): Option[string] =
-        if v != "":
-          result = some v & "!"
-        else:
-          result = none(string)
-
-      check(some(1).flatMap(maybeToString).flatMap(maybeExclaim) == some("1!"))
-      check(some(0).flatMap(maybeToString).flatMap(maybeExclaim) == none(string))
-
     test "SomePointer":
       var intref: ref int
       check(option(intref).isNone)
@@ -503,4 +345,3 @@ when isMainModule:
     test "Ref type with overloaded `==`":
       let p = some(RefPerson.new())
       check p.isSome
-